Create UI and Server files for final project
Created the Ui and server files for our Shiny application. We currently have a dropbox that allows the user to select a building of their choice and a checkbox to select up to three explanatory variables. If no varibles are selected, the graph will not display. The graph is currently not synced up to the explanatory variables.